Fix stale MTP stdout acceptance summary and dedupe MTP discovery
- RunMtpApplicationSurfacesPerTestStandardOutput asserted the old 4-test
  summary (2,1,1); the shared MtpMSTestProject asset gained a 5th test
  (RunSettingsEnvironmentVariableIsInjected), so the run is (3,1,1).
  This is a semantic collision between #16283 and #16284 that only
  surfaces when both are on main. Update the expected summary.
- Address review feedback: extract the near-identical MTP discovery
  routine from MtpProxyDiscoveryManager and MtpProxyExecutionManager into
  a single MtpClientHelpers.DiscoverSourceTests helper so the two paths
  cannot drift (timeouts, cancellation, sentinel drain, env-var policy).

Copilot-Session: 61d9a1c3-3c9f-4dab-8ba3-5baba8d55ed6
