+++ src/Microsoft.TestPlatform.CrossPlatEngine/Client/MTP/MtpClientHelpers.cs
@@ -4,9 +4,12 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
+using System.Threading;
 
 using Jsonite;
 
+using Microsoft.VisualStudio.TestPlatform.ObjectModel;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 
 namespace Microsoft.VisualStudio.TestPlatform.CrossPlatEngine.Client.MTP;
@@ -52,6 +55,75 @@ internal static class MtpClientHelpers
         }
 
         return TimeSpan.FromSeconds(90);
+    }
+
+    /// <summary>
+    /// Runs a single MTP discovery pass against <paramref name="source"/> and returns the discovered tests.
+    /// Shared by the discovery proxy (which forwards the tests to the discovery handler) and the execution
+    /// proxy (which resolves a <c>/TestCaseFilter</c> against the returned set). Discovery is started with no
+    /// environment variables so execution-only data-collector profiler variables are never injected.
+    /// </summary>
+    /// <param name="source">The MTP application to discover.</param>
+    /// <param name="logHandler">Receives log messages produced by the MTP application.</param>
+    /// <param name="cancellationToken">Cancels the discovery pass.</param>
+    public static List<TestCase> DiscoverSourceTests(
+        string source,
+        Action<TestMessageLevel, string?> logHandler,
+        CancellationToken cancellationToken)
+    {
+        var discovered = new List<TestCase>();
+        using var completed = new ManualResetEventSlim(false);
+
+        using var connection = new MtpServerConnection();
+        connection.LogReceived += (level, message) => logHandler(MapLevel(level), message);
+        connection.TestNodesUpdated += parameters =>
+        {
+            if (IsCompletionSentinel(parameters))
+            {
+                completed.Set();
+                return;
+            }
+
+            foreach (JsonObject node in EnumerateNodes(parameters))
+            {
+                if (MtpTestNodeConverter.IsActionNode(node))
+                {
+                    lock (discovered)
+                    {
+                        discovered.Add(MtpTestNodeConverter.ToTestCase(node, source));
+                    }
+                }
+            }
+        };
+
+        connection.Start(source, environmentVariables: null, GetConnectionTimeout());
+        connection.InvokeAsync(MtpConstants.InitializeMethod, InitializeParameters(), cancellationToken).GetAwaiter().GetResult();
+
+        var runId = Guid.NewGuid();
+        var discoverTask = connection.InvokeAsync(
+            MtpConstants.DiscoverTestsMethod,
+            new Dictionary<string, object?> { [MtpConstants.RunIdParameter] = runId.ToString() },
+            cancellationToken);
+
+        // The DiscoverTests response indicates the server finished discovery. Because messages arrive on a
+        // single ordered stream that we read sequentially, every node notification sent before the response
+        // has already been dispatched, so 'discovered' is complete once the response returns. Wait briefly
+        // for the trailing completion sentinel (honoring cancellation) purely to drain it; not observing it
+        // does not invalidate the discovered set.
+        discoverTask.GetAwaiter().GetResult();
+        if (!completed.Wait(TimeSpan.FromSeconds(3), cancellationToken))
+        {
+            EqtTrace.Warning(
+                "MtpClientHelpers.DiscoverSourceTests: discovery for '{0}' did not signal the completion sentinel within the drain window; results reflect the nodes received so far.",
+                source);
+        }
+
+        connection.SendNotification(MtpConstants.ExitMethod, null);
+
+        lock (discovered)
+        {
+            return discovered.ToList();
+        }
     }
 
     private static int GetCurrentProcessId()
